chore: re-order components
