+++ src/views/overview/components/backing/AssetBreakdown.tsx
@@ -1,27 +1,17 @@
+import AsteriskIcon from 'components/icons/AsteriskIcon'
+import CirclesIcon from 'components/icons/CirclesIcon'
 import CollaterizationIcon from 'components/icons/CollaterizationIcon'
+import LayersIcon from 'components/icons/LayersIcon'
+import RiskIcon from 'components/icons/RiskIcon'
 import TabMenu from 'components/tab-menu'
 import { useCallback, useMemo, useState } from 'react'
+import { ChevronDown } from 'react-feather'
 import { Box, Card, Flex, Text } from 'theme-ui'
+import BackingOverview from './BackingOverview'
 import CollateralExposure from './CollateralExposure'
-import TokenExposure from './TokenExposure'
 import PlatformExposure from './PlatformExposure'
 import Risks from './Risks'
-import CirclesIcon from 'components/icons/CirclesIcon'
-import LayersIcon from 'components/icons/LayersIcon'
-import CollateralsChart from './CollateralsChart'
-import { atom, useAtom, useAtomValue } from 'jotai'
-import {
-  estimatedApyAtom,
-  rTokenAtom,
-  rTokenBackingDistributionAtom,
-  rTokenPriceAtom,
-} from 'state/atoms'
-import EarnNavIcon from 'components/icons/EarnNavIcon'
-import { formatCurrency } from 'utils'
-import { Trans } from '@lingui/macro'
-import CircleIcon from 'components/icons/CircleIcon'
-import EarnIcon from 'components/icons/EarnIcon'
-import RiskIcon from 'components/icons/RiskIcon'
+import TokenExposure from './TokenExposure'
 
 const tabComponents = {
   collaterals: CollateralExposure,
@@ -29,13 +19,6 @@ const tabComponents = {
   platforms: PlatformExposure,
   risks: Risks,
 }
-
-// const tabCharts = {
-//   collaterals: CollateralExposure,
-//   tokens: TokenExposure,
-//   platforms: PlatformExposure,
-//   risks: Risks,
-// }
 
 const Menu = ({
   current,
@@ -65,77 +48,6 @@ const Menu = ({
   return <TabMenu active={current} items={items} onMenuChange={onChange} />
 }
 
-// TODO: TARGET PEG PRICE (ETH+)
-const backingOverviewAtom = atom((get) => {
-  const rToken = get(rTokenAtom)
-  const distribution = get(rTokenBackingDistributionAtom)
-  const price = get(rTokenPriceAtom)
-  const apys = get(estimatedApyAtom)
-
-  return {
-    symbol: rToken?.symbol ?? '',
-    backing: distribution?.backing ?? 0,
-    staked: distribution?.staked ?? 0,
-    yield: apys.basket,
-    price: price,
-  }
-})
-
-const BackingOverviewContainer = ({ current }: { current: string }) => {
-  const data = useAtomValue(backingOverviewAtom)
-
-  return (
-    <Card
-      variant="inner"
-      mr={3}
-      p={4}
-      mt={[3, 3, 3, 0]}
-      sx={{
-        flexDirection: 'column',
-        width: ['100%', '100%', '100%', 256],
-        height: ['fit-content', 'fit-content', 'fit-content', 360],
-        flexShrink: '0',
-        fontSize: [1, 2, 2, 1],
-        // justifyContent: 'center',
-      }}
-    >
-      <Box variant="layout.verticalAlign">
-        <EarnNavIcon fontSize={16} />
-        <Text ml="2">1 {data.symbol}</Text>
-        <Text ml="auto" variant="strong">
-          ${formatCurrency(data?.price)}
-        </Text>
-      </Box>
-      <Box mt="2" variant="layout.verticalAlign">
-        <CircleIcon color="currentColor" />
-        <Text ml="2">
-          <Trans>Backing</Trans>
-        </Text>
-        <Text ml="auto" variant="strong">
-          {data.backing.toFixed(0)}%
-        </Text>
-      </Box>
-      <Box sx={{ display: ['none', 'none', 'none', 'block'] }}>
-        <CollateralsChart />
-      </Box>
-      <Box mt={[2, 2, 2, 0]} variant="layout.verticalAlign">
-        <CollaterizationIcon fontSize={16} />
-        <Text ml="2">Staked RSR</Text>
-        <Text ml="auto" variant="strong">
-          {data.staked.toFixed(0)}%
-        </Text>
-      </Box>
-      <Box mt="2" variant="layout.verticalAlign">
-        <EarnIcon color="currentColor" />
-        <Text ml="2">Blended Yield</Text>
-        <Text ml="auto" variant="strong">
-          {data.yield.toFixed(2)}%
-        </Text>
-      </Box>
-    </Card>
-  )
-}
-
 const AssetBreakdown = () => {
   const [current, setCurrent] = useState('collaterals')
 
@@ -152,6 +64,11 @@ const AssetBreakdown = () => {
     <Card>
       <Box variant="layout.verticalAlign">
         <Menu current={current} onChange={handleChange} />
+        <Box variant="layout.verticalAlign" ml="auto">
+          <AsteriskIcon />
+          <Text mx="2">What if there’s a de-peg?</Text>
+          <ChevronDown size={16} />
+        </Box>
       </Box>
       <Flex
         mt={3}
@@ -159,7 +76,7 @@ const AssetBreakdown = () => {
           flexWrap: ['wrap-reverse', 'wrap-reverse', 'wrap-reverse', 'nowrap'],
         }}
       >
-        <BackingOverviewContainer current={current} />
+        <BackingOverview current={current} />
         <Current />
       </Flex>
     </Card>
