+++ src/views/overview/components/backing/TokenExposure.tsx
@@ -1,7 +1,20 @@
-import { Box } from 'theme-ui'
+import { Box, Card } from 'theme-ui'
 
 const TokenExposure = () => {
-  return <Box></Box>
+  return (
+    <Card variant="inner">
+      <Box
+        variant="layout.verticalAlign"
+        p={4}
+        sx={{ borderBottom: '1px solid', borderColor: 'border' }}
+      >
+        {/* <RiskIcon />
+        <Text ml="2" mr="auto" variant="bold" sx={{ fontSize: 3 }}>
+          <Trans>Other Risks</Trans>
+        </Text> */}
+      </Box>
+    </Card>
+  )
 }
 
 export default TokenExposure
