+++ src/views/overview/index.tsx
@@ -5,7 +5,7 @@ import mixpanel from 'mixpanel-browser'
 import { useEffect } from 'react'
 import { selectedRTokenAtom } from 'state/atoms'
 import { Divider } from 'theme-ui'
-import Hero from './components/Hero'
+import Hero from './components/hero'
 import RTokenEarn from './components/RTokenEarn'
 import RTokenTransactions from './components/RTokenTransactions'
 import Backing from './components/backing'
